Apollo-Project-Orchestrator-Local/Apollo-Project-Orchestrator-Local/backend/requirements/dev.txt
# requirements/dev.txt - Dependências de desenvolvimento
# Inclui todas as dependências base
-r base.txt

# =============================================================================
# DESENVOLVIMENTO E DEBUG
# =============================================================================
flask-debugtoolbar==0.13.1

# =============================================================================
# TESTES
# =============================================================================
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-xdist==3.3.1
factory-boy==3.3.0
coverage==7.3.2

# =============================================================================
# LINTING E FORMATAÇÃO
# =============================================================================
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
pylint==2.17.7

# =============================================================================
# SEGURANÇA
# =============================================================================
bandit==1.7.5
safety==2.3.4

# =============================================================================
# DOCUMENTAÇÃO
# =============================================================================
sphinx==7.2.6
flask-swagger-ui==4.11.1
apispec==6.3.0
apispec-webframeworks==0.5.2

# =============================================================================
# PRE-COMMIT HOOKS
# =============================================================================
pre-commit==3.4.0

# =============================================================================
# DEBUGGING E PROFILING
# =============================================================================
ipdb==0.13.13
memory-profiler==0.61.0

# =============================================================================
# LOAD TESTING
# =============================================================================
locust==2.17.0